+++ src/array.h
@@ -24,24 +24,24 @@
 /* ------------------------------------------------------------ */
 typedef struct array_s
 {
-  int size;
-  int count;
+  size_t size;
+  size_t count;
   void **v;
 } array_t;
 typedef array_t *array_pt;
 
 /* ------------------------------------------------------------ */
-array_pt array_new(unsigned int);
+array_pt array_new(size_t);
 void array_fill(array_pt, void *);
-array_pt array_new_fill(unsigned int, void *);
+array_pt array_new_fill(size_t, void *);
 void array_free(array_pt);
 void array_clear(array_pt);
 array_pt array_clone(array_pt);
-unsigned int array_add(array_pt, void *);
-unsigned int array_add_unique(array_pt, void *);
+size_t array_add(array_pt, void *);
+size_t array_add_unique(array_pt, void *);
 void array_delete_item(array_pt, void *);
 void array_delete_duplicates(array_pt);
-void *array_delete_index(array_pt, unsigned int);
+void *array_delete_index(array_pt, size_t);
 void array_trim(array_pt);
 void array_filter_with(array_pt, int (*)(void *, void *), void *);
 void array_filter(array_pt, int (*)(void *));
@@ -50,11 +50,11 @@ void array_map1(array_pt, void (*)(void *, void *), void *);
 void array_map2(array_pt, void (*)(void *, void *, void *), void *, void *);
 #define array_map_with(a, b, c) array_map1(a, b, c)
 
-int array_size(array_pt a);
-int array_count(array_pt a);
-void *array_get(array_pt a, int i);
+size_t array_size(array_pt a);
+size_t array_count(array_pt a);
+void *array_get(array_pt a, size_t i);
 
-void *array_set(array_pt, int, void *);
+void *array_set(array_pt, size_t, void *);
 
 /* ------------------------------------------------------------ */
 #endif
